Improve event handling
This improves event handling in the main player fragment by remembering
and replaying the most recently received event when the fragment's
views are recreated. The views are recreated when the fragment comes
back to the foreground after having been on the backstack, and events
are received infrequently enough (once per second, typically) that
it was possible for the user to see an out-of-date UI (in the state
it was in when the user moved it to the background). By storing and
replaying the most recent event, the UI is not seen in a stale state.